Add feedback card skeleton
